# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PICG API Documentation",
        "description": "This documentation covers all API endpoints for the PICG project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/skill": {
            "get": {
                "tags": [
                    "Skill"
                ],
                "summary": "Get Skills by User ID",
                "operationId": "getSkills",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of the user whose skills to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of skills",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "skill_name": {
                                                "type": "string",
                                                "example": "Laravel"
                                            },
                                            "level": {
                                                "type": "string",
                                                "example": "Intermediate"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No skills found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Skill"
                ],
                "summary": "Create a new skill",
                "operationId": "storeSkill",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "name",
                                    "skill_proficiency",
                                    "description"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Laravel"
                                    },
                                    "skill_proficiency": {
                                        "type": "string",
                                        "enum": [
                                            "Beginner",
                                            "Intermediate",
                                            "Expert"
                                        ],
                                        "example": "Intermediate"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Experience building RESTful APIs"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Skill created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Laravel"
                                                },
                                                "skill_proficiency": {
                                                    "type": "string",
                                                    "example": "Intermediate"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Experience building RESTful APIs"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-11T10:45:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-11T10:45:00Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Skill Create Successfully!."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/skill/show": {
            "get": {
                "tags": [
                    "Skill"
                ],
                "summary": "Get skill details by ID",
                "operationId": "showSkill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID of the skill to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Laravel"
                                        },
                                        "skill_proficiency": {
                                            "type": "string",
                                            "example": "Expert"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Developed APIs and admin dashboards."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Skill not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/skill/update": {
            "post": {
                "tags": [
                    "Skill"
                ],
                "summary": "Update a skill",
                "operationId": "updateSkill",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "skill_id"
                                ],
                                "properties": {
                                    "skill_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Laravel"
                                    },
                                    "skill_proficiency": {
                                        "type": "string",
                                        "enum": [
                                            "Beginner",
                                            "Intermediate",
                                            "Expert"
                                        ],
                                        "example": "Expert"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Experience with Laravel framework."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skill updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/skill/delete": {
            "delete": {
                "tags": [
                    "Skill"
                ],
                "summary": "Delete a skill by ID",
                "operationId": "deleteSkill",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skill deleted successfully"
                    },
                    "404": {
                        "description": "Skill not found"
                    },
                    "500": {
                        "description": "Failed to delete skill"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}